requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.9.1
    # via django
black==25.1.0
    # via -r requirements.in
click==8.2.1
    # via black
coverage[toml]==7.9.2
    # via pytest-cov
django==5.2.4
    # via
    #   -r requirements.in
    #   django-model-utils
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   model-bakery
django-environ==0.12.0
    # via -r requirements.in
django-model-utils==5.0.0
    # via -r requirements.in
djangorestframework==3.16.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.5.0
    # via -r requirements.in
flake8==7.3.0
    # via -r requirements.in
iniconfig==2.1.0
    # via pytest
isort==6.0.1
    # via -r requirements.in
mccabe==0.7.0
    # via flake8
model-bakery==1.20.5
    # via -r requirements.in
mypy-extensions==1.1.0
    # via black
packaging==25.0
    # via
    #   black
    #   pytest
pathspec==0.12.1
    # via black
platformdirs==4.3.8
    # via black
pluggy==1.6.0
    # via
    #   pytest
    #   pytest-cov
psycopg2-binary==2.9.10
    # via -r requirements.in
pycodestyle==2.14.0
    # via flake8
pyflakes==3.4.0
    # via flake8
pygments==2.19.2
    # via pytest
pyjwt==2.9.0
    # via djangorestframework-simplejwt
pytest==8.4.1
    # via
    #   -r requirements.in
    #   pytest-cov
    #   pytest-django
pytest-cov==6.2.1
    # via -r requirements.in
pytest-django==4.11.1
    # via -r requirements.in
sqlparse==0.5.3
    # via django
